Fixed loop waiting for EB. Removed dependency on custom DNS.
The loop in config_db.sh waiting for the WebAPI to respond a 200 code
does not work if the custom Route53 DNS option was selected, as the DNS
records are not yet created. Creating them before this template does not
fix the issue the curl command on a DNS name that does not resolve to a
valid CNAME (case while the EB environment is in creation process),
return a void result, maing the while expression fail.
The applied solution checks for the status of the EB environment to be
Ready and its health Green.
To avoid an endless loop a WaitCondition on the template is added and
signaled by the User Data script after config_db.sh is completed.