Modified BCrypt hash generation to make it accepted by Atlas
Previously the htpasswd command was used, which generates $2y version of
the hash algorithm. The script replaced $y with $a as the Spring BCrytp
library used by WebAPI does not recognize the $2y version. This hack
does not work and the generated value itself is different between $2y
and $2a versions.
This change uses a Python script that generates a valid $2a version
hash.