Deleted temp instance and wait conditions after deploy
The previous behavior of terminating the instance after completing the
bootstrapping makes subsequent update to the templates fail, as the
instance resource does not exist anymore.
A better bootstrapping could be achieved using CodePipeline/Codebuild,
that could be reused for veersion updates and would not create this kind
of operational conflic. (Future enhancement)