Added scheduling of ASG instances stop and start
Configures an ASG action to set Max, Min and Desired capacity to zero,
and another to restore the selected values. Both based on a recurrence
schedule. IF the schedule is set to NEVER, the action is configured to
take place starting so far in the future it will never happen