Removed second R installation that is ignored and breaks shiny-server
The R installation in /mnt/r-stuff/R-latest is not used in the
installation process. In fact, the script itself installs the
dependencies in the system installation. But the configuration of this
secondary installation in PATH environment variable through /etc/profile
breaks shiny-server that fails to load package 'digest'. The reason is
that shiny-server launches worker processes as the 'shiny' user through
a login session that as such, uses the PATH in /etc/profile and hence
the secondary R installation that does not have any of the dependencies.
Shiny-Server itself works because it is launched from init, which has
its own PATH and does not load /etc/profile.