Added 240 sec Tomcat timeout to WebAPI ebextensions 00HIPAA.config
The rest of timeout values (ELB and Apache PHP) were already set to 240
seconds, but the tomcat connection Timeout was at the default 20000
miliseconds.
This change is untested as it would require a new deployment. Could be
tested by manually adding the change to the deployed application version
(to be done)